+++ src/extension_instance.ml
@@ -3,6 +3,9 @@ open Import
 type t =
   { mutable sandbox : Sandbox.t
   ; mutable repl : Terminal_sandbox.t option
+  ; mutable ocaml_version : Ocaml_version.t option
+        (** assumption: it must be set before initializing the language server;
+            the lang server initialization needs the ocaml version *)
   ; mutable lsp_client : (LanguageClient.t * Ocaml_lsp.t) option
   ; sandbox_info : StatusBarItem.t
   ; ast_editor_state : Ast_editor_state.t
@@ -16,92 +19,103 @@ let ocaml_lsp t = Option.map ~f:snd t.lsp_client
 
 let lsp_client t = t.lsp_client
 
-let client_options () =
-  let documentSelector =
-    LanguageClient.DocumentSelector.
-      [| language "ocaml"
-       ; language "ocaml.interface"
-       ; language "ocaml.ocamllex"
-       ; language "ocaml.menhir"
-       ; language "reason"
-      |]
-  in
-  let (lazy outputChannel) = Output.language_server_output_channel in
-  let revealOutputChannelOn = LanguageClient.RevealOutputChannelOn.Never in
-  LanguageClient.ClientOptions.create ~outputChannel ~revealOutputChannelOn
-    ~documentSelector ()
-
-let server_options sandbox =
-  let command = Sandbox.get_command sandbox "ocamllsp" [] in
-  Cmd.log command;
-  let env =
-    let extra_env_vars =
-      Settings.server_extraEnv () |> Option.value ~default:Interop.Dict.empty
-    in
-    Interop.Dict.union (fun _k _v1 v2 -> Some v2) Process.Env.env extra_env_vars
-  in
-  match command with
-  | Shell command ->
-    let options = LanguageClient.ExecutableOptions.create ~env ~shell:true () in
-    LanguageClient.Executable.create ~command ~options ()
-  | Spawn { bin; args } ->
-    let command = Path.to_string bin in
-    let options =
-      LanguageClient.ExecutableOptions.create ~env ~shell:false ()
-    in
-    LanguageClient.Executable.create ~command ~args ~options ()
+let ocaml_version_exn t = Option.value_exn t.ocaml_version
 
 let stop_server t =
   Option.iter t.lsp_client ~f:(fun (client, _) ->
       t.lsp_client <- None;
       LanguageClient.stop client)
 
-let start_language_server t =
-  stop_server t;
-  let res =
-    let open Promise.Result.Syntax in
-    let* () =
-      let cmd = Sandbox.get_command t.sandbox "ocamllsp" [ "--version" ] in
-      let open Promise.Syntax in
-      let+ (res : Node.ChildProcess.return) = Cmd.run cmd in
-      if res.exitCode = 0 then
-        Ok ()
-      else
-        Error "Sandbox initialisation failed: ocaml-lsp-server is not installed"
+module Language_server_init : sig
+  val start_language_server : t -> unit Promise.t
+end = struct
+  let client_options () =
+    let documentSelector =
+      LanguageClient.DocumentSelector.
+        [| language "ocaml"
+         ; language "ocaml.interface"
+         ; language "ocaml.ocamllex"
+         ; language "ocaml.menhir"
+         ; language "reason"
+        |]
     in
-    let serverOptions = server_options t.sandbox in
-    let clientOptions = client_options () in
-    let client =
-      LanguageClient.make ~id:"ocaml" ~name:"OCaml Platform VS Code extension"
-        ~serverOptions ~clientOptions ()
-    in
-    LanguageClient.start client;
+    let (lazy outputChannel) = Output.language_server_output_channel in
+    let revealOutputChannelOn = LanguageClient.RevealOutputChannelOn.Never in
+    LanguageClient.ClientOptions.create ~outputChannel ~revealOutputChannelOn
+      ~documentSelector ()
 
+  let server_options sandbox =
+    let command = Sandbox.get_command sandbox "ocamllsp" [] in
+    Cmd.log command;
+    let env =
+      let extra_env_vars =
+        Settings.server_extraEnv () |> Option.value ~default:Interop.Dict.empty
+      in
+      Interop.Dict.union
+        (fun _k _v1 v2 -> Some v2)
+        Process.Env.env extra_env_vars
+    in
+    match command with
+    | Shell command ->
+      let options =
+        LanguageClient.ExecutableOptions.create ~env ~shell:true ()
+      in
+      LanguageClient.Executable.create ~command ~options ()
+    | Spawn { bin; args } ->
+      let command = Path.to_string bin in
+      let options =
+        LanguageClient.ExecutableOptions.create ~env ~shell:false ()
+      in
+      LanguageClient.Executable.create ~command ~args ~options ()
+
+  let check_ocaml_lsp_available sandbox =
+    let ocaml_lsp_version sandbox =
+      Sandbox.get_command sandbox "ocamllsp" [ "--version" ]
+    in
+    Cmd.output (ocaml_lsp_version sandbox)
+    |> Promise.Result.fold
+         ~ok:(fun (_ : string) -> ())
+         ~error:(fun (_ : string) ->
+           "Sandbox initialization failed: `ocaml-lsp-server` is not installed \
+            in the current sandbox.")
+
+  let start_language_server t =
+    stop_server t;
     let open Promise.Syntax in
-    let+ initialize_result = LanguageClient.readyInitializeResult client in
-    let ocaml_lsp = Ocaml_lsp.of_initialize_result initialize_result in
-    t.lsp_client <- Some (client, ocaml_lsp);
-    if
-      (not (Ocaml_lsp.has_interface_specific_lang_id ocaml_lsp))
-      || (not (Ocaml_lsp.can_handle_switch_impl_intf ocaml_lsp))
-      || not (Ocaml_lsp.can_handle_infer_intf ocaml_lsp)
-      (* TODO: switch to ocaml-lsp version based approach Using
-         [initializeResult] of [LanguageClient] we can get ocaml-lsp's version.
-         We can use versions instead of capabilities to suggest the user to
-         update their ocaml-lsp. *)
-    then
-      show_message `Warn
-        "The installed version of ocamllsp is out of date. Some features may \
-         be unavailable or degraded in functionality: switching between \
-         implementation and interface files, functionality in mli sources. \
-         Consider updating ocamllsp.";
-    Ok ()
-  in
-  let open Promise.Syntax in
-  let+ res = res in
-  match res with
-  | Ok () -> ()
-  | Error s -> show_message `Error "Error starting server: %s" s
+    let+ res =
+      let open Promise.Result.Syntax in
+      let* () = check_ocaml_lsp_available t.sandbox in
+      let client =
+        let serverOptions = server_options t.sandbox in
+        let clientOptions = client_options () in
+        LanguageClient.make ~id:"ocaml" ~name:"OCaml Platform VS Code extension"
+          ~serverOptions ~clientOptions ()
+      in
+      LanguageClient.start client;
+      let open Promise.Syntax in
+      let+ initialize_result = LanguageClient.ready_initialize_result client in
+      let ocaml_lsp = Ocaml_lsp.of_initialize_result initialize_result in
+      t.lsp_client <- Some (client, ocaml_lsp);
+      (match
+         Ocaml_lsp.is_version_up_to_date ocaml_lsp (ocaml_version_exn t)
+       with
+      | Ok is_up_to_date ->
+        if not is_up_to_date then
+          show_message `Warn
+            "The installed version of `ocamllsp` is out of date. You may be \
+             missing out on cool features or important bug fixes. Consider \
+             upgrading the package `ocaml-lsp-server`."
+      | Error (`Msg m) -> show_message `Warn "%s" m);
+      Ok ()
+    in
+    match res with
+    | Ok () -> ()
+    | Error s ->
+      show_message `Error
+        "An error occurred starting the language server `ocamllsp`. %s" s
+end
+
+include Language_server_init
 
 module Sandbox_info : sig
   val make : Sandbox.t -> StatusBarItem.t
@@ -131,7 +145,13 @@ let make () =
   let sandbox = Sandbox.Global in
   let sandbox_info = Sandbox_info.make sandbox in
   let ast_editor_state = Ast_editor_state.make () in
-  { sandbox; lsp_client = None; sandbox_info; repl = None; ast_editor_state }
+  { sandbox
+  ; lsp_client = None
+  ; sandbox_info
+  ; repl = None
+  ; ocaml_version = None
+  ; ast_editor_state
+  }
 
 let set_sandbox t new_sandbox =
   Sandbox_info.update t.sandbox_info ~new_sandbox;
@@ -150,6 +170,57 @@ let repl t = t.repl
 let set_repl t repl = t.repl <- Some repl
 
 let close_repl t = t.repl <- None
+
+let update_ocaml_info t =
+  let open Promise.Syntax in
+  let+ ocaml_version =
+    let+ r =
+      Sandbox.get_command t.sandbox "ocamlc" [ "--version" ] |> Cmd.output
+    in
+    match r with
+    | Ok v ->
+      Ocaml_version.of_string v
+      |> Result.map_error ~f:(function m ->
+             `Unable_to_parse_version (`Version v, m))
+    | Error e ->
+      log_chan ~section:"Ocaml.version_semver" `Warn
+        "Error running `ocamlc --version`: %s" e;
+      Error `Ocamlc_missing
+  in
+  match ocaml_version with
+  | Ok ocaml_version -> t.ocaml_version <- Some ocaml_version
+  | Error e -> (
+    (* [t.ocaml_version <- None] because we don't want [t.ocaml_version] to be
+       left over from a previous sandbox, which successfully set it *)
+    t.ocaml_version <- None;
+    match e with
+    | `Unable_to_parse_version (`Version v, `Msg msg) ->
+      show_message `Error
+        "OCaml bytecode compiler `ocamlc` version could not be parsed. \
+         Version: %s. Error %s"
+        v msg
+    | `Ocamlc_missing ->
+      let (_ : unit Promise.t) =
+        let+ maybe_choice =
+          Window.showErrorMessage
+            ~message:
+              "OCaml bytecode compiler `ocamlc` was not found in the current \
+               sandbox. Do you have OCaml installed in the current sandbox?"
+            ~choices:
+              [ ( "Pick another sandbox"
+                , fun () ->
+                    let (_ : Ojs.t option Promise.t) =
+                      Vscode.Commands.executeCommand
+                        ~command:Extension_consts.Commands.select_sandbox
+                        ~args:[]
+                    in
+                    () )
+              ]
+            ()
+        in
+        Option.iter maybe_choice ~f:(fun f -> f ())
+      in
+      ())
 
 let open_terminal sandbox =
   let terminal = Terminal_sandbox.create sandbox in