+++ src/extension_instance.mli
@@ -14,6 +14,8 @@ val ocaml_lsp : t -> Ocaml_lsp.t option
 
 val lsp_client : t -> (LanguageClient.t * Ocaml_lsp.t) option
 
+val ocaml_version_exn : t -> Ocaml_version.t
+
 val start_language_server : t -> unit Promise.t
 
 val open_terminal : Sandbox.t -> unit
@@ -25,5 +27,7 @@ val repl : t -> Terminal_sandbox.t option
 val set_repl : t -> Terminal.t -> unit
 
 val close_repl : t -> unit
+
+val update_ocaml_info : t -> unit Promise.t
 
 val ast_editor_state : t -> Ast_editor_state.t